+++ projects/_01_ingest/cqc_api/fargate/validate_cqc_locations_4_full_clean.py
@@ -5,9 +5,9 @@ import pointblank as pb
 from polars_utils import utils
 from polars_utils.expressions import str_length_cols
 from polars_utils.validation import actions as vl
+from polars_utils.validation.actions import add_list_column_validation_check_flags
 from polars_utils.validation.constants import GLOBAL_ACTIONS, GLOBAL_THRESHOLDS
 from projects._01_ingest.cqc_api.utils.validate_cqc_locations import (
-    add_list_column_validation_check_flags,
     get_expected_row_count_for_validation_full_clean,
 )
 from utils.column_names.cleaned_data_files.cqc_location_cleaned import (
